Fix that chroot is not properly tested for user and group
When testing if the required user and/ or group exist,
the check was performed on the main filesystem and not
in chroot if one was provided.

This was causing failures if the required user or group
were present on the main filesystem, but not in chroot.

With this change, the test for user and group is performed
in chroot if one is provided, or in the main filesystem
otherwise.